Estandarizar deleteProperty y añadir las funcionalidades para admins

+++ src/main/java/com/habitame/api/property/service/PropertyService.java
@@ -3,20 +3,20 @@ package com.habitame.api.property.service;
 import com.habitame.api.auth.security.SecurityUtils;
 import com.habitame.api.city.service.CityService;
 import com.habitame.api.common.exception.ResourceNotFoundException;
+import com.habitame.api.common.exception.UnauthorizedException;
 import com.habitame.api.common.mapper.PropertyMapper;
 import com.habitame.api.common.wrapper.PageResponse;
 import com.habitame.api.property.dto.*;
 import com.habitame.api.property.entity.PropertyEntity;
 import com.habitame.api.property.entity.PropertyStatus;
-import com.habitame.api.property.repository.PropertyListProjection;
 import com.habitame.api.property.repository.PropertyRepository;
 import com.habitame.api.propertyReview.service.PropertyReviewService;
+import com.habitame.api.user.entity.Role;
+import com.habitame.api.user.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.apache.catalina.security.SecurityUtil;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -28,6 +28,7 @@ public class PropertyService {
     private final PropertyRepository propertyRepository;
     private final CityService cityService;
     private final PropertyReviewService propertyReviewService;
+    private final UserService userService;
 
     public PageResponse<PropertyPublicResponse> findPublicProperties(Pageable pageable) {
 
@@ -57,7 +58,7 @@ public class PropertyService {
         return propertyRepository.findById(propertyId).orElseThrow(() -> new ResourceNotFoundException("Property not found: " + propertyId));
     }
 
-    public PageResponse<PropertyOwnerResponse> findAllByOwner(Pageable pageable){
+    public PageResponse<PropertyOwnerResponse> findAllByOwner(Pageable pageable) {
         Integer ownerId = SecurityUtils.getCurrentUserId();
 
         Page<PropertyEntity> page = propertyRepository.findAllByOwnerId(ownerId, pageable).orElseThrow(() -> new ResourceNotFoundException("Owner not found: " + ownerId));
@@ -100,9 +101,9 @@ public class PropertyService {
     public PropertyOwnerDetailResponse updateOwnerProperty(Integer propertyId, @Valid PropertyOwnerRequest request) {
         PropertyEntity propertyEntity = propertyRepository.findByIdAndOwnerId(SecurityUtils.getCurrentUserId(), propertyId).orElseThrow(() -> new ResourceNotFoundException("Property not found: " + propertyId));
         PropertyEntity propertyToUpdate = PropertyMapper.updateProperty(propertyEntity, request, cityService.findEntityById(request.getCityId()));
-        if(!propertyEntity.getTitle().equals(request.getTitle()) ||
-            !propertyEntity.getDescription().equals(request.getDescription()) ||
-            !propertyEntity.getAddress().equals(request.getAddress())) {
+        if (!propertyEntity.getTitle().equals(request.getTitle()) ||
+                !propertyEntity.getDescription().equals(request.getDescription()) ||
+                !propertyEntity.getAddress().equals(request.getAddress())) {
             propertyToUpdate.setStatus(PropertyStatus.IN_REVIEW);
             propertyReviewService.addReview(propertyEntity);
         }
@@ -111,9 +112,13 @@ public class PropertyService {
     }
 
     @Transactional
-    public void deleteOwnerProperty(Integer idProperty) {
-        PropertyEntity propertyEntity = propertyRepository.findByIdAndOwnerId(SecurityUtils.getCurrentUserId(), idProperty).orElseThrow(() -> new ResourceNotFoundException("Property not found: " + idProperty));
-        propertyRepository.delete(propertyEntity);
+    public void deleteProperty(Integer idProperty) {
+        PropertyEntity propertyEntity = propertyRepository.findById(idProperty).orElseThrow(() -> new ResourceNotFoundException("Property not found: " + idProperty));
+        if (propertyEntity.getOwner() != SecurityUtils.getCurrentUser() && SecurityUtils.getCurrentUser().getRole() != Role.ADMIN) {
+            throw new UnauthorizedException("No tienes permiso para eliminar");
+        } else {
+            propertyRepository.delete(propertyEntity);
+        }
     }
 
     public PageResponse<PropertyAdminResponse> findAll(Pageable pageable) {
@@ -130,5 +135,23 @@ public class PropertyService {
                 page.getTotalElements(),
                 page.getTotalPages()
         );
+    }
+
+    public PropertyAdminDetailResponse findById(Integer idProperty) {
+        PropertyEntity propertyEntity = propertyRepository.findById(idProperty).orElseThrow(() -> new ResourceNotFoundException("Property not found: " + idProperty));
+        return PropertyMapper.toAdminDetailResponse(propertyEntity);
+    }
+
+    public PropertyAdminResponse addAdminProperty(PropertyAdminRequest request) {
+        PropertyEntity propertyEntity = PropertyMapper.adminToEntity(request, userService.findById(request.getOwnerId()), cityService.findEntityById(request.getCityId()));
+        return PropertyMapper.toAdminResponse(propertyRepository.save(propertyEntity));
+    }
+
+    public PropertyAdminDetailResponse updateAdminProperty(Integer idProperty, PropertyAdminRequest request) {
+        PropertyEntity propertyEntity = propertyRepository.findById(idProperty).orElseThrow(() -> new ResourceNotFoundException("Property not found: " + idProperty));
+        PropertyEntity propertyToUpdate = PropertyMapper.updateAdminProperty(propertyEntity, request, cityService.findEntityById(request.getCityId()), userService.findById(request.getOwnerId()));
+        propertyToUpdate.setUpdatedBy(SecurityUtils.getCurrentUser());
+        propertyRepository.save(propertyToUpdate);
+        return PropertyMapper.toAdminDetailResponse(propertyToUpdate);
     }
 }
